requirements.txt
python-telegram-bot
feedparser
deep-translator
beautifulsoup4